Refactor out tests module and remove empty subtree
